handle failures; authorize invitations

+++ app/controllers/memberships_controller.rb
@@ -1,18 +1,20 @@
 class MembershipsController < ApplicationController
   def create
     invitation = Invitation.find_by(league_id: params[:league_id], user_id: current_user.id)
+
     if invitation
-      Membership.create(league_id: params[:league_id], user_id: current_user.id)
-      invitation.destroy
-      flash[:notice] = 'Success'
+      membership = Membership.create(league_id: params[:league_id], user_id: current_user.id)
+
+      if membership
+        invitation.destroy
+        flash[:notice] = 'Invitation Accepted'
+      else
+        flash[:error] = 'Invitation could not be accepted'
+      end
     else
-      flash[:alert] = 'Nice Try'
+      flash[:error] = 'Invitation could not be accepted'
     end
 
     redirect_to user_leagues_path(user_id: current_user.id)
-  end
-
-  def destroy
-
   end
 end
